Add update page and remove useless pages
