+++ content/updates/example.md
@@ -1,0 +1,5 @@
+# An example
+::: meta-data type="New" ee-only="false" link-to-doc="../articles/what-is-the-pim-marketplace.html"
+:::
+
+This is an example
