--- content/updates/index.json
@@ -1,3 +1,0 @@
-{
-
-}
