--- content/whats-new/1.7-news.json
@@ -1,51 +1,0 @@
-[
-  {
-    "title":"Teamwork Assistant",
-    "areas": ["Productivity", "Collaboration"],
-    "editions": ["EE"],
-    "since-version": "1.7",
-    "link": "/pim/v2/articles/what-is-a-project.html"
-  },
-  {
-    "title":"Products API 🔌",
-    "areas": ["Governance", "Connectivity"],
-    "editions": ["CE","GE","EE"],
-    "since-version": "1.7",
-    "link": "https://api.akeneo.com/api-reference.html#Products"
-  },
-  {
-    "title":"Categories API 🔌",
-    "areas": ["Governance", "Connectivity"],
-    "editions": ["CE","GE","EE"],
-    "since-version": "1.7",
-    "link": "https://api.akeneo.com/api-reference.html#Category"
-  },
-  {
-    "title":"Families API 🔌",
-    "areas": ["Governance", "Connectivity"],
-    "editions": ["CE","GE","EE"],
-    "since-version": "1.7",
-    "link": "https://api.akeneo.com/api-reference.html#Family"
-  },
-  {
-    "title":"Attributes API 🔌",
-    "areas": ["Governance", "Connectivity"],
-    "editions": ["CE","GE","EE"],
-    "since-version": "1.7",
-    "link": "https://api.akeneo.com/api-reference.html#Attribute"
-  },
-  {
-    "title":"Channels API 🔌",
-    "areas": ["Governance", "Connectivity"],
-    "editions": ["CE","GE","EE"],
-    "since-version": "1.7",
-    "link": "https://api.akeneo.com/api-reference.html#Channel"
-  },
-  {
-    "title":"Locales API 🔌",
-    "areas": ["Governance", "Connectivity"],
-    "editions": ["CE","GE","EE"],
-    "since-version": "1.7",
-    "link": "https://api.akeneo.com/api-reference.html#Locale"
-  }
-]
