--- content/whats-new/2.0-news.json
@@ -1,71 +1,0 @@
-[
-  {
-    "title":"New modeling for products with variants",
-    "areas": ["Governance", "Productivity"],
-    "editions": ["CE","GE", "EE"],
-    "since-version": "2.0",
-    "link": "/pim/v2/articles/what-about-products-variants.html"
-  },
-  {
-    "title":"Product models API 🔌",
-    "areas": ["Governance", "Connectivity"],
-    "editions": ["CE","GE", "EE"],
-    "since-version": "2.0",
-    "link": "https://api.akeneo.com/api-reference-index.html#Productmodel"
-  },
-  {
-    "title":"Family variants API 🔌",
-    "areas": ["Governance", "Connectivity"],
-    "editions": ["CE","GE", "EE"],
-    "since-version": "2.0",
-    "link": "https://api.akeneo.com/api-reference-index.html#Familyvariant"
-  },
-  {
-    "title":"Permissions on product data on imports/exports 🔒",
-    "areas": ["Connectivity"],
-    "editions": ["EE"],
-    "since-version": "2.0",
-    "link": "/pim/v2/articles/imports.html#with-the-user-groups-rights-ee-only"
-  },
-  {
-    "title":"Brand new UX/UI 🖌️",
-    "areas": ["Productivity"],
-    "editions": ["CE","GE", "EE"],
-    "since-version": "2.0"
-  },
-  {
-    "title":"Attribute groups API 🔌",
-    "areas": ["Governance", "Connectivity"],
-    "editions": ["CE","GE", "EE"],
-    "since-version": "2.0",
-    "link": "https://api.akeneo.com/api-reference-index.html#Attributegroup"
-  },
-  {
-    "title":"Association types API 🔌",
-    "areas": ["Governance", "Connectivity"],
-    "editions": ["CE","GE", "EE"],
-    "since-version": "2.0",
-    "link": "https://api.akeneo.com/api-reference-index.html#Associationtype"
-  },
-  {
-    "title":"Currencies API 🔌",
-    "areas": ["Governance", "Connectivity"],
-    "editions": ["CE","GE", "EE"],
-    "since-version": "2.0",
-    "link": "https://api.akeneo.com/api-reference-index.html#Currency"
-  },
-  {
-    "title":"Published products API 🔌",
-    "areas": ["Connectivity"],
-    "editions": ["EE"],
-    "since-version": "2.0",
-    "link": "https://api.akeneo.com/api-reference-index.html#Publishedproduct"
-  },
-  {
-    "title":"Product draft management via API 🔌",
-    "areas": ["Connectivity","Collaboration"],
-    "editions": ["EE"],
-    "since-version": "2.0",
-    "link": "https://api.akeneo.com/api-reference.html#get_draft__code_"
-  }
-]
