--- content/whats-new/2.1-news.json
@@ -1,23 +1,0 @@
-[
-  {
-    "title":"Search bar in the product grid 🔍",
-    "areas": ["Productivity"],
-    "editions": ["CE","GE", "EE"],
-    "since-version": "2.1",
-    "link": "/pim/v2/articles/products-grid.html#search-by-label-and-identifier-since-21"
-  },
-  {
-    "title":"Bulk categorization on product models 📂",
-    "areas": ["Productivity"],
-    "editions": ["CE","GE", "EE"],
-    "since-version": "2.1",
-    "link": "/pim/v2/articles/product-mass-actions.html#categorize-move-or-remove-products"
-  },
-  {
-    "title":"Gallery view in the product grid 🖼️",
-    "areas": ["Productivity"],
-    "editions": ["CE","GE", "EE"],
-    "since-version": "2.1",
-    "link": "/pim/v2/articles/products-grid.html#product-grid-display-since-21"
-  }
-]
