--- content/whats-new/2.2-news.json
@@ -1,23 +1,0 @@
-[
-  {
-    "title":"Rules compatible with product models 🤖",
-    "areas": ["Automation", "Productivity"],
-    "editions": ["EE"],
-    "since-version": "2.2",
-    "link": "/pim/v2/articles/manage-your-rules.html#execute-rules"
-  },
-  {
-    "title":"Search filters in the proposal grid 🔍",
-    "areas": ["Productivity","Collaboration"],
-    "editions": ["EE"],
-    "since-version": "2.2",
-    "link": "/pim/v2/articles/review-products-proposals.html#filter-the-proposals-since-22"
-  },
-  {
-    "title":"Bulk addition of product attribute values",
-    "areas": ["Productivity"],
-    "editions": ["CE","GE","EE"],
-    "since-version": "2.2",
-    "link": "/pim/v2/articles/product-mass-actions.html#add-new-values-to-attributes-since-22"
-  }
-]
