--- content/whats-new/2.3-news.json
@@ -1,30 +1,0 @@
-[
-  {
-    "title":"Associations compatible with product models 🔗",
-    "areas": ["Governance"],
-    "editions": ["CE","GE","EE"],
-    "since-version": "2.3",
-    "link": "/pim/v2/articles/product-mass-actions.html#associate-products-since-21"
-  },
-  {
-    "title":"Proposals compatible with product models",
-    "areas": ["Productivity","Collaboration"],
-    "editions": ["EE"],
-    "since-version": "2.3",
-    "link": "/pim/v2/articles/proposals-workflow.html#overview"
-  },
-  {
-    "title":"Parent change for sub-product model or variant product",
-    "areas": ["Governance","Productivity"],
-    "editions": ["CE","GE","EE"],
-    "since-version": "2.3",
-    "link": "/pim/v2/articles/product-mass-actions.html#change-the-parent-product-model-since-23"
-  },
-  {
-    "title":"Assets upload in the product form 🖼️",
-    "areas": ["Productivity"],
-    "editions": ["EE"],
-    "since-version": "2.3",
-    "link": "/pim/v2/articles/work-on-a-product.html#upload-assets-directly-from-the-product-form"
-  }
-]
