--- content/whats-new/3.0-news.json
@@ -1,49 +1,0 @@
-[
-  {
-    "title":"Reference entities to manage product-related data",
-    "areas": ["Governance"],
-    "editions": ["EE"],
-    "since-version": "3.0",
-    "link": "/pim/v3/articles/what-about-reference-entities.html"
-  },
-  {
-    "title":"Completeness on reference entities",
-    "areas": ["Data quality"],
-    "editions": ["EE"],
-    "since-version": "3.0",
-    "link": "/pim/v3/articles/manage-reference-entities.html#define-the-completeness"
-  },
-  {
-    "title":"Permissions on reference entities 🔒",
-    "areas": ["Governance","Collaboration"],
-    "editions": ["EE"],
-    "since-version": "3.0",
-    "link": "/pim/v3/articles/access-rights-on-entities.html"
-  },
-  {
-    "title":"Link a reference entity to another one 🔗",
-    "areas": ["Governance"],
-    "editions": ["EE"],
-    "since-version": "3.0",
-    "link": "/pim/v3/articles/enrich-your-reference-entity-records.html#link-a-record-to-another-record"
-  },
-  {
-    "title":"Reference entities API 🔌",
-    "areas": ["Connectivity"],
-    "editions": ["EE"],
-    "since-version": "3.0",
-    "link": "https://api.akeneo.com/api-reference-index.html#Referenceentities"
-  },
-  {
-    "title":"Faster product grid 🚀",
-    "areas": ["Productivity","Scalability"],
-    "editions": ["CE","GE", "EE"],
-    "since-version": "3.0"
-  },
-  {
-    "title":"Faster dashboard 🚀",
-    "areas": ["Reporting","Scalability"],
-    "editions": ["CE","GE", "EE"],
-    "since-version": "3.0"
-  }
-]
