--- content/whats-new/3.1-news.json
@@ -1,9 +1,0 @@
-[
-  {
-    "title":"New filters in the reference entity grid",
-    "areas": ["Productivity"],
-    "editions": ["EE"],
-    "since-version": "3.1",
-    "link": "/pim/v3/articles/enrich-your-reference-entity-records.html#filter-on-options-and-reference-entity-links-since-31"
-  }
-]
