--- content/whats-new/3.2-news.json
@@ -1,29 +1,0 @@
-[
-  {
-    "title":"Number attribute type available in reference entities",
-    "areas": ["Governance"],
-    "editions": ["EE"],
-    "since-version": "3.2",
-    "link": "/pim/v3/articles/manage-reference-entities.html#add-an-attribute"
-  },
-  {
-    "title":"Display of products linked to a reference entity record 🔗",
-    "areas": ["Productivity"],
-    "editions": ["EE"],
-    "since-version": "3.2",
-    "link": "/pim/v3/articles/enrich-your-reference-entity-records.html#display-the-products-linked-to-a-record-since-32"
-  },
-  {
-    "title":"Proposals sent for approval in bulk",
-    "areas": ["Productivity","Collaboration"],
-    "editions": ["EE"],
-    "since-version": "3.2",
-    "link": "/pim/v3/articles/proposals-workflow.html#via-a-bulk-actions"
-  },
-  {
-    "title":"Improved API product export performance 🚀",
-    "areas": ["Connectivity", "Scalability"],
-    "editions": ["CE","GE", "EE"],
-    "since-version": "3.2"
-  }
-]
