--- content/whats-new/4.0-news.json
@@ -1,170 +1,0 @@
-[
-  {
-    "title":"Asset modeling thanks to asset families 🖼️",
-    "areas": ["Governance"],
-    "editions": ["EE"],
-    "since-version": "4.0",
-    "link": "/pim/v4/updates/asset-manager.html#more-flexibility-to-manage-your-assets"
-  },
-  {
-    "title":"Asset enrichment thanks to asset attributes 🖼️",
-    "areas": ["Governance","Data quality"],
-    "editions": ["EE"],
-    "since-version": "4.0",
-    "link": "/pim/v4/updates/asset-manager.html#more-flexibility-to-manage-your-assets"
-  },
-  {
-    "title":"Contextualized asset data by locale and channel",
-    "areas": ["Governance","Data quality"],
-    "editions": ["EE"],
-    "since-version": "4.0",
-    "link": "/pim/v4/updates/asset-manager.html#contextualize-your-asset-information"
-  },
-  {
-    "title":"Easy 3rd-party asset retrieval 🖼️",
-    "areas": ["Governance","Connectivity"],
-    "editions": ["EE"],
-    "since-version": "4.0",
-    "link": "/pim/v4/updates/asset-manager.html#more-flexibility-to-manage-your-assets"
-  },
-  {
-    "title":"Completeness on assets 🖼️",
-    "areas": ["Data quality"],
-    "editions": ["EE"],
-    "since-version": "4.0",
-    "link": "/pim/v4/updates/asset-manager.html#permissions-completeness"
-  },
-  {
-    "title":"Flexible asset transformations 🤖",
-    "areas": ["Productivity","Automation"],
-    "editions": ["EE"],
-    "since-version": "4.0",
-    "link": "/pim/v4/updates/asset-manager.html#improved-asset-transformation-rules"
-  },
-  {
-    "title":"Improved mass asset upload",
-    "areas": ["Productivity"],
-    "editions": ["EE"],
-    "since-version": "4.0",
-    "link": "/pim/v4/updates/asset-manager.html#mass-upload"
-  },
-  {
-    "title":"Asset automatically linked to products 🤖",
-    "areas": ["Productivity","Automation"],
-    "editions": ["EE"],
-    "since-version": "4.0",
-    "link": "/pim/v4/updates/asset-manager.html#automatically-link-assets-to-products-with-product-link-rules"
-  },
-  {
-    "title":"Assets API 🔌",
-    "areas": ["Connectivity"],
-    "editions": ["EE"],
-    "since-version": "4.0",
-    "link": "https://api.akeneo.com/api-reference-index.html#AssetManager"
-  },
-  {
-    "title":"Data quality grades on products 💯",
-    "areas": ["Data quality", "Reporting"],
-    "editions": ["EE"],
-    "since-version": "4.0",
-    "link": "/pim/v4/updates/data-quality-insights.html#monitor-your-data-quality-in-a-dashboard"
-  },
-  {
-    "title":"Data quality monitoring dashboard 📊",
-    "areas": ["Data quality", "Reporting"],
-    "editions": ["EE"],
-    "since-version": "4.0",
-    "link": "/pim/v4/updates/data-quality-insights.html#monitor-your-data-quality-in-a-dashboard"
-  },
-  {
-    "title":"Spell-check on text area and text attributes in French 🇫🇷",
-    "areas": ["Data quality"],
-    "editions": ["EE"],
-    "since-version": "4.0",
-    "link": "/pim/v4/updates/data-quality-insights.html#how-does-it-work"
-  },
-  {
-    "title":"Spell-check on text area and text attributes in English 🇬🇧",
-    "areas": ["Data quality"],
-    "editions": ["EE"],
-    "since-version": "4.0",
-    "link": "/pim/v4/updates/data-quality-insights.html#how-does-it-work"
-  },
-  {
-    "title":"Spell-check on text area and text attributes in German 🇩🇪",
-    "areas": ["Data quality"],
-    "editions": ["EE"],
-    "since-version": "4.0",
-    "link": "/pim/v4/updates/data-quality-insights.html#how-does-it-work"
-  },
-  {
-    "title":"Spell-check on text area and text attributes in Spanish 🇪🇸",
-    "areas": ["Data quality"],
-    "editions": ["EE"],
-    "since-version": "4.0",
-    "link": "/pim/v4/updates/data-quality-insights.html#how-does-it-work"
-  },
-  {
-    "title":"API connections management 🔌",
-    "areas": ["Connectivity"],
-    "editions": ["CE","GE","EE"],
-    "since-version": "4.0",
-    "link": "/pim/v4/articles/manage-your-connections.html"
-  },
-  {
-    "title":"Streamlined API connectors configuration 🔌",
-    "areas": ["Connectivity"],
-    "editions": ["CE","GE","EE"],
-    "since-version": "4.0",
-    "link": "/pim/v4/updates/connections.html#a-new-streamlined-way-to-configure-connectors-from-within-the-pim"
-  },
-  {
-    "title":"API product import monitoring 📊",
-    "areas": ["Connectivity","Reporting"],
-    "editions": ["CE","GE","EE"],
-    "since-version": "4.0",
-    "link": "/pim/v4/updates/connections.html#a-new-dashboard-to-monitor-all-your-connections"
-  },
-  {
-    "title":"2x to 4x faster API product import 🚀",
-    "areas": ["Connectivity","Scalability"],
-    "editions": ["CE","GE","EE"],
-    "since-version": "4.0",
-    "link": "/pim/v4/updates/performance.html#better-import-performance"
-  },
-  {
-    "title":"3x to 6x faster API product model import 🚀",
-    "areas": ["Connectivity","Scalability"],
-    "editions": ["CE","GE","EE"],
-    "since-version": "4.0",
-    "link": "/pim/v4/updates/performance.html#better-import-performance"
-  },
-  {
-    "title":"1.5x to 9x faster CSV product import 🚀",
-    "areas": ["Connectivity","Scalability"],
-    "editions": ["CE","GE","EE"],
-    "since-version": "4.0",
-    "link": "/pim/v4/updates/performance.html#better-import-performance"
-  },
-  {
-    "title":"6x to 12x faster CSV product model import 🚀",
-    "areas": ["Connectivity","Scalability"],
-    "editions": ["CE","GE","EE"],
-    "since-version": "4.0",
-    "link": "/pim/v4/updates/performance.html#better-import-performance"
-  },
-  {
-    "title":"30% faster API product export 🚀",
-    "areas": ["Connectivity","Scalability"],
-    "editions": ["CE","GE","EE"],
-    "since-version": "4.0",
-    "link": "/pim/v4/updates/performance.html#better-export-performance"
-  },
-  {
-    "title":"CSV product export performance twice faster 🚀",
-    "areas": ["Connectivity","Scalability"],
-    "editions": ["CE","GE","EE"],
-    "since-version": "4.0",
-    "link": "/pim/v4/updates/performance.html#better-export-performance"
-  }
-]
