--- content/whats-new/6.0-news.json
@@ -1,107 +1,0 @@
-[
-  {
-    "title":"Table Attribute",
-    "areas": ["Governance"],
-    "editions": ["EE", "GE"],
-    "since-version": "6.0",
-    "link": "/pim/serenity/updates/2021-10.html#manage-your-multidimensional-data-in-akeneo-pim"
-  },
-  {
-    "title":"Tailored exports 🪡",
-    "areas": ["Collaboration", "Productivity"],
-    "editions": ["EE", "GE"],
-    "since-version": "6.0",
-    "link": "/pim/serenity/updates/2021-08.html#advanced-product-export-capabilities"
-  },
-  {
-    "title":"Asset mass edition 🖼️",
-    "areas": ["Productivity"],
-    "editions": ["EE"],
-    "since-version": "6.0",
-    "link": "/pim/serenity/updates/2021-03.html#mass-edit-your-assets"
-  },
-  {
-    "title":"Attribute guidelines",
-    "areas": ["Governance"],
-    "editions": ["EE","CE", "GE"],
-    "since-version": "6.0",
-    "link": "/pim/serenity/updates/2021-03.html#add-guidelines-to-your-attributes"
-  },
-  {
-    "title":"Asset mass deletion 🖼️",
-    "areas": ["Productivity"],
-    "editions": ["EE"],
-    "since-version": "6.0",
-    "link": "/pim/serenity/updates/2021-02.html#mass-delete-your-assets"
-  },
-  {
-    "title":"User duplication 👤",
-    "areas": ["Productivity"],
-    "editions": ["EE","CE", "GE"],
-    "since-version": "6.0",
-    "link": "/pim/serenity/updates/2021-03.html#duplicate-a-user"
-  },
-  {
-    "title":"User import and export in Excel or CSV 👤",
-    "areas": ["Productivity"],
-    "editions": ["EE","CE", "GE"],
-    "since-version": "6.0",
-    "link": "/pim/serenity/updates/2021-03.html#import-and-export-your-users-roles-and-groups"
-  },
-  {
-    "title":"Events API logs",
-    "areas": ["Productivity", "Connectivity"],
-    "editions": ["EE","CE", "GE"],
-    "since-version": "6.0",
-    "link": "/pim/serenity/updates/2021-03.html#browse-your-events-api-logs"
-  },
-  {
-    "title":"Spell-check dictionaries management 📗",
-    "areas": ["Data quality","Governance"],
-    "editions": ["EE"],
-    "since-version": "6.0",
-    "link": "/pim/serenity/updates/2021-02.html#manage-your-spell-check-dictionaries"
-  },
-  {
-    "title":"Product quality score export in Excel or CSV files",
-    "areas": ["Data quality","Collaboration"],
-    "editions": ["EE","CE", "GE"],
-    "since-version": "6.0",
-    "link": "/pim/serenity/updates/2021-02.html#filter-and-export-your-product-quality-score"
-  },
-  {
-    "title":"Product completeness export via API 🔌",
-    "areas": ["Connectivity"],
-    "editions": ["EE","CE", "GE"],
-    "since-version": "6.0",
-    "link": "/pim/serenity/updates/2021-05.html#retrieve-product-completeness-through-api"
-  },
-  {
-    "title":"Product quality score export via API 🔌",
-    "areas": ["Data quality","Connectivity"],
-    "editions": ["EE","CE", "GE"],
-    "since-version": "6.0",
-    "link": "/pim/serenity/updates/2021-02.html#new-api-filter-on-product-quality-score"
-  },
-  {
-    "title":"Reference entity record mass deletion",
-    "areas": ["Productivity"],
-    "editions": ["EE"],
-    "since-version": "6.0",
-    "link": "/pim/serenity/updates/2021-04.html#mass-delete-your-reference-entity-records"
-  },
-  {
-    "title":"Redesigned proposal screen",
-    "areas": ["Collaboration","Productivity"],
-    "editions": ["EE"],
-    "since-version": "6.0",
-    "link": "/pim/serenity/updates/2021-04.html#review-your-proposals-more-efficiently"
-  },
-  {
-    "title":"10 million assets per asset family 🖼️",
-    "areas": ["Governance"],
-    "editions": ["EE"],
-    "since-version": "6.0",
-    "link": "/pim/serenity/updates/2021-02.html#up-to-10-millions-of-assets-per-asset-family"
-  }
-]
