--- content/whats-new/Serenity-news.json
@@ -1,2 +1,0 @@
-[
-]
