--- content/whats-next/whats-next-mid-term.json
@@ -1,12 +1,0 @@
-[
-  {
-    "title": "Multiple Product Identifiers",
-    "description": "Oftentimes Akeneo customers need to use more than one product identifier for a single product. With this release, you can use several identifiers per product to suit your needs, like GTIN and SKU, as well as internal identifiers.",
-    "areas": ["Governance"]
-  },
-  {
-    "title": "Simultaneous Work in Akeneo PIM",
-    "description": "With many teams collaborating on the enrichment of products within Akeneo PIM, they need a way to know if someone else is also working on that product. With Simulatneous Work in Akeneo PIM, users will not only see when someone else is working on the same product, but they can also work on the same product at the same time without overwriting each other’s product values.",
-    "areas": ["Productivity"]
-  }
-]
