--- content/whats-next/whats-next-short-term.json
@@ -1,23 +1,0 @@
-[
-  {
-    "title": "Akeneo PIM - Product Identifier Generator",
-    "description": "Improve product lifecycle management by enabling unique product identifier generation within Akeneo PIM. Generate IDs based on nomenclature, like combining attributes, or use reference entities to do so. Take advantange of this increased flexibility in Akeneo SaaS versions.",
-    "areas": ["Governance"],
-    "img": "./img/IDGeneratorMedium.png",
-    "editions": ["EE","GE"]
-  },
-  {
-    "title": "Enriched Categories",
-    "description": "Creating a great customer experience means providing additional context at every step of their buying journey. Therefore Enriched Categories will give Akeneo customers the ability to enrich not just products, but categories as well, including channel- and locale-specific attributes too!",
-    "img": "./img/enrichedcats1Large.png",
-    "img": "./img/enrichedcats2Large.png",
-    "areas": ["Governance"]
-  },
-  {
-    "title": "Akeneo Apps - Catalogs with attribute mapping",
-    "description": "Easing app development and configuration is key for Akeneo partners and developers. With attribute mapping for catalogs developers can reduce the TCO and improve the time to deployment, since native mapping capabilities inside Akeneo PIM will enable users to connect and then map their PIM attributes to the app targets.",
-    "img": "./img/Mapping-enriched.png",
-    "areas": ["Connectivity"],
-    "editions": ["EE","GE"]
-  }
-]
