--- src/monthly-updates-index.handlebars
@@ -1,35 +1,0 @@
-{{!-- The index of the monthly updates --}}
-{{#> layout }}
-{{> navbar-with-search }}
-{{> whats-new-button}}
-    <div class="container">
-        <div class="row">
-            <div class="col-xs-12">
-                <p class="breadcrumb text-muted"><a href="/">Home</a> / <a href="/pim/v7">PIM help center</a> / What's new in Akeneo PIM 7.0?</p>
-                <h1>What's new in Akeneo PIM 7.0?</h1>
-                <p><em>Akeneo PIM 7.0 will help you reach the next level with new features & improvements. Akeneo PIM 7.0 is a Long-Term Support (LTS) version. It will be supported for <b>18 months</b> until the end of September 2024.</em></p>
-            </div>
-        </div>
-        <div class="row">
-            {{#each yearlyUpdates}}
-                <div class="col-xs-2 col-md-1">
-                    <h2>{{year}}</h2>
-                </div>
-                <div class="col-xs-10 col-md-11">
-                    <hr style="margin: 58px 0 32px 0;"></hr>
-                </div>
-                {{#each monthlyUpdates as |monthlyUpdate id|}}
-                    <div class="col-xs-5 col-md-4 col-lg-3">
-                        <img class="img-responsive center-block" src="../img/{{img}}" style="margin-top: 50px;"/>
-                    </div>
-                    <div class="col-xs-12 col-sm-7 col-md-8 col-lg-9">
-                        <h3>{{title}}</h3>
-                        <p><em>{{description}}</em></p>
-                        <a href="{{id}}.html" class="btn btn-primary btn-ghost btn-sm" style="margin-bottom: 20px;">Read more</a>
-                    </div>
-                <div class="col-xs-12"></div>
-                {{/each}}
-            {{/each}}
-        </div>
-    </div>
-{{/layout}}
