--- src/monthly-updates.handlebars
@@ -1,15 +1,0 @@
-{{!-- Shared template for all the monthly update pages --}}
-{{#> layout }}
-{{> navbar-with-search }}
-{{> whats-new-button}}
-    <div class="container updates">
-        <div class="row">
-            <div class="col-xs-12">
-                <p class="breadcrumb text-muted"><a href="/">Home</a> / <a href="/pim/v7">PIM help center</a> / <a href="/pim/v7/updates/index.html"> What's new in Akeneo PIM Serenity Enterprise & Growth editions?</a> / {{title}}</p>
-                <h1>{{title}}</h1>
-                <p><em>{{description}}</em></p>
-                <div class="row">{{{mainContent}}}</div>
-            </div>
-        </div>
-    </div>
-{{/layout}}
