+++ src/partials/whats-new-button.handlebars
@@ -1,6 +1,6 @@
- <a href="/pim/v7/whats-new.html">
+ <a href="/pim/serenity/whats-new.html">
     <div class="whats-new-btn">
-        <img src="/pim/v7/img/illus-whats-new.svg">
+        <img src="/pim/{{majorVersion}}/img/illus-whats-new.svg">
         <em>Discover our features per version</em>
     </div>
 </a>
