+++ src/updates.handlebars
@@ -1,0 +1,17 @@
+{{!-- Shared template for all the release note update pages --}}
+{{#> layout }}
+{{> navbar-with-search }}
+{{> whats-new-button}}
+    <div class="container">
+        <div class="row">
+            <div class="col-xs-12">
+                <p class="breadcrumb text-muted"><a href="/">Home</a> / <a href="/pim/{{majorVersion}}">PIM help center</a> / <a href="/pim/{{majorVersion}}/updates/index.html"> What's new in Akeneo PIM 7.0?</a></p>
+                <h1>What's new in Akeneo PIM 7.0 ?</h1>
+                <p><em>Akeneo PIM 7.0 will help you reach the next level with new features & improvements. Akeneo PIM 7.0 is a Long-Term Support (LTS) version. It will be supported for <b>18 months</b> until the end of September 2024.</em></p>
+            </div>
+        </div>
+        <div class="row updates">
+            {{{mainContent}}}
+        </div>
+    </div>
+{{/layout}}
